# apps/trench/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_root","parameters":[],"responses":{"200":{"description":""}}}},"/events":{"post":{"operationId":"EventsController_handleEvent","summary":"Create one or more events","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventsDTO"}}}},"responses":{"200":{"description":"The events have been successfully created."}},"tags":["events"],"security":[{"bearer":[]}]}},"/queries":{"post":{"operationId":"QueriesController_executeQueries","summary":"Execute queries","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueriesDTO"}}}},"responses":{"200":{"description":"The queries have been successfully executed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedQueryResponse"}}}}}}},"/webhooks":{"get":{"operationId":"WebhooksController_getWebhooks","summary":"Get all webhooks","parameters":[],"responses":{"200":{"description":"The webhooks have been successfully retrieved.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedWebhookResponse"}}}}},"security":[{"bearer":[]}]},"post":{"operationId":"WebhooksController_createWebhook","summary":"Create a webhook","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookDTO"}}}},"responses":{"200":{"description":"The webhook has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Webhook"}}}}},"security":[{"bearer":[]}]}},"/webhooks/{uuid}":{"delete":{"operationId":"WebhooksController_deleteWebhook","summary":"Delete a webhook","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The webhook has been successfully deleted."}},"security":[{"bearer":[]}]}}},"info":{"title":"trench API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"EventDTO":{"type":"object","properties":{"anonymousId":{"type":"string","description":"The anonymous ID of the user. See [anonymousId](https://segment.com/docs/connections/spec/common/#anonymousid).","example":"guest-1234567890"},"event":{"type":"string","description":"The event name. Used only with `type: \"track\"`.","example":"UserSignedUp"},"timestamp":{"type":"string","description":"The timestamp of the event. Autogenerated if not provided.","example":"2021-01-01T00:00:00Z"},"type":{"type":"string","description":"The type of the event. Possible values: `track`, `identify`, `group`.","example":"track"},"userId":{"type":"string","description":"The ID of the user.","example":"user-1234567890"},"groupId":{"type":"string","description":"The ID of the group.","example":"group-1234567890"},"properties":{"type":"object","description":"The properties of the event. Only used with `type: \"track\"`.","example":{"page":"/home","referrer":"https://www.google.com"}},"traits":{"type":"object","description":"The traits of the user or group.","example":{"email":"john.doe@example.com","name":"John Doe"}}}},"EventsDTO":{"type":"object","properties":{"events":{"description":"The events to handle.","type":"array","items":{"$ref":"#/components/schemas/EventDTO"}}},"required":["events"]},"QueriesDTO":{"type":"object","properties":{"queries":{"description":"The queries to execute.","example":["SELECT COUNT(*) FROM events WHERE event = \"UserSignedUp\""],"type":"array","items":{"type":"string"}}},"required":["queries"]},"PaginatedQueryResponse":{"type":"object","properties":{"results":{"description":"The results of the queries, returned in the same order as the queries.","example":[{"results":[{"count":3485241}]}],"type":"array","items":{"type":"object"}}},"required":["results"]},"Webhook":{"type":"object","properties":{"uuid":{"type":"string","description":"The UUID of the webhook. Automatically generated.","example":"123e4567-e89b-12d3-a456-426614174000"},"url":{"type":"string","description":"The URL that the webhook will send events to.","example":"https://your-webhook-url.com"},"enableBatching":{"type":"boolean","description":"Whether to enable batching for the webhook.","example":true},"createdAt":{"format":"date-time","type":"string","description":"The date and time the webhook was created.","example":"2021-01-01T00:00:00.000Z"},"eventTypes":{"description":"The event types that the webhook will send. Use `*` to match all event types.","example":["track","identify"],"type":"array","items":{"type":"string"}},"eventNames":{"description":"The event names that the webhook will send. Use `*` to match all event names.","example":["UserSignedUp","UserLoggedIn"],"type":"array","items":{"type":"string"}}},"required":["uuid","url","enableBatching","createdAt","eventTypes","eventNames"]},"PaginatedWebhookResponse":{"type":"object","properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/Webhook"}}},"required":["results"]},"WebhookDTO":{"type":"object","properties":{"url":{"type":"string","description":"The URL that the webhook will send events to.","example":"https://your-webhook-url.com"},"enableBatching":{"type":"boolean","description":"Whether to enable batching for the webhook. Defaults to `false`.","example":true},"eventTypes":{"description":"The event types that the webhook will send. Defaults to `[\"*\"] (all event types)`.","example":["track","identify"],"type":"array","items":{"type":"string"}},"eventNames":{"description":"The event names that the webhook will send. Defaults to `[\"*\"] (all event names)`.","example":["UserSignedUp","UserLoggedIn"],"type":"array","items":{"type":"string"}}},"required":["url"]}}}}